@@ -8,8 +8,7 @@
             "name": "Fredrik Liljegren"
         },
         {
-            "name": "Sören Jensen",
-            "email": "soren@abicart.se"
+            "name": "Sören Jensen"
         }
     ],
     "autoload": {
@@ -23,11 +22,14 @@
         }
     },
     "require": {
-        "php": "^7.2 | ^8.0",
-        "psr/log": "^1 | ^2 | ^3"
+        "php": "^7.4 | ^8.0",
+        "phrity/net-uri": "^1.0",
+        "phrity/util-errorhandler": "^1.0",
+        "psr/log": "^1.0 | ^2.0 | ^3.0",
+        "psr/http-message": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0|^9.0",
+        "phpunit/phpunit": "^9.0",
         "php-coveralls/php-coveralls": "^2.0",
         "squizlabs/php_codesniffer": "^3.5"
     }
