@@ -7,9 +7,16 @@
  * License text: https://raw.githubusercontent.com/Textalk/websocket-php/master/COPYING
  */
 
-namespace WebSocket\Message;
+namespace WebSocket;
 
-class Text extends Message
+trait OpcodeTrait
 {
-    protected $opcode = 'text';
+    private static $opcodes = [
+        'continuation' => 0,
+        'text'         => 1,
+        'binary'       => 2,
+        'close'        => 8,
+        'ping'         => 9,
+        'pong'         => 10,
+    ];
 }
